+++ zammadoo/__init__.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: UTF-8 -*-
 
-__version__ = "0.1.0"
+__version__ = "0.2.0.dev0"
 
 from .client import LOG, APIException
 from .client import Client as Client  # pylint: disable=useless-import-alias