+++ HISTORY.rst
@@ -5,8 +5,7 @@ History
 future release
 --------------
 * **[breaking change]** Client parameter: instead ``username`` and ``password`` now ``http_auth=(username, password)`` is provided
-* add Client.server_version()
 
 0.1.0 (2023-10-08)
------------
+------------------
 * initial release