+++ HISTORY.rst
@@ -8,6 +8,7 @@ future release
 --------------
 * **breaking change**: :class:`client.Client` now uses ``http_auth=(username, password)`` as authentication parameter
 * **new feature**: :meth:`tickets.Ticket.history` method
+* **fix**: resource items are now cached when using iteration
 
 0.1.0 (2023-10-08)
 ------------------