+++ HISTORY.rst
@@ -6,8 +6,9 @@ History
 
 future release
 --------------
-* **new feature**
+* **added features**
 
+    * extend support for Python 3.8
     * added :meth:`tickets.Ticket.history` method
     * added ``weburl`` property for :class:`client.Client`, :class:`tickets.Ticket`,
       :class:`users.User` and :class:`organizations.Organization`
@@ -22,7 +23,7 @@ future release
 * **fixes**
 
     * resource items are now cached when using iteration
-    * fromisoformat conversion in Python 3.9 supporting Zulu offset format
+    * fromisoformat conversion in Python <3.10 supporting Zulu offset format
 
 0.1.0 (2023-10-08)
 ------------------